feat(assert): add assert state verification to Run resource
This commit lets one to assert an entire k8s resource by just
specifying the same. This state must have apiVersion & kind to
be considered for assert. The optional fields that are used
as identifier fields are name, namespace, labels, & annotations.

All other fields if specified in state will be asserted with the
objects found in k8s cluster.

Signed-off-by: AmitKumarDas <amit.das@mayadata.io>