feat(testing): initial integration test implementation
This commit uses container image to run integration tests. The
tests are automated and run on a K3s installation. Integration
tests are triggered as a make target.

e.g. make integration-test-suite

This approach provides the code coverage report in suite's run
logs.

This change fixes a labelling bug where resources that are not
supposed to be labelled are labelled.

GitHub actions has been modifed to accommodate integration testing.

Signed-off-by: AmitKumarDas <amit.das@mayadata.io>