chore(testing): update readme w.r.t programmatic vs declarative testing
This commit makes changes to clearly indicate programmatic as well
declarative way to test D-operators' Kubernetes controllers.

It is important to note that the same model can be followed by other
Operators repo if needed. This model has worked fine so far. It makes
use of GitHub VM as launching machine & same VM used as a single node
K3s cluster to run the tests (both declarative as well as programmatic).

Signed-off-by: AmitKumarDas <amit.das@mayadata.io>